Fix tile animations with no end-of-animation which moved upwards forever
It's a little known trick that by filling up the whole animation pattern there
will be no end-of-animation and hence it never resets to 0 offset, and you can
use to to create long animations, eg a whole column of tiles. One or two of
Mogri's games do this. But it only worked if you went down or right, not up or
left.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13591 7d344553-34f0-0310-a9b1-970ce8f1c3a2