Oops I reimplemented the slice detail menu unintentionally
All I wanted was a list of slice properties... then I noticed all the data I
wanted was already in the slice detail menu, so I set about transforming the
code to extract the data. But it was so much easier to keep it as imperative
code rather than a giant array initialiser. Rather similar to a menu refresh
function... and then it dawned on me that EdKit was already an ideal API for
describing editable data. Of course, I had to make sure it actually
compiles. And if it compiles, might as well execute it and inspect the results
look correct. And when they don't, what am I meant to do but get it functional?

However I did delete some stuff in the copy initially because I didn't think
I was trying to create an editor.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13758 7d344553-34f0-0310-a9b1-970ce8f1c3a2