sliceedit: convert the detail menu refresh function to the method of a SliceDetailMenu class and switch to BasicMenuItem
The reason I did this was to fix an annoying graphical bug where the header
lines in the menu weren't highlighted when selected, because they used fgtag.
To fix that I used BasicMenuItem.col instead to set their text colour. But I
didn't want to change all the "a_append menu()" lines and cause some huge
conflicts in some branches I have, so I converted the whole thing to a method so
sliceed_header could be a method. Ugh.

Of course after doing that I found it still too hard to selected header lines,
so used a darker colour.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13764 7d344553-34f0-0310-a9b1-970ce8f1c3a2