Don't call CreateDocument or init_intern_string at the module level
In r13749 I ensured that type_table's are statically initialised so that vectors
and HashTables can constructed at module level, because global_reload_doc was
created at the module level in common.rbas, but that had to be reverted in
r13774, after which the engine runs into an error on startup. So instead,
global_reload_doc is now created in a call at the top of game/custom.bas, and
init_intern_string gets called from lowlevel_init (also called from the top of
main()), which is much saner.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13777 7d344553-34f0-0310-a9b1-970ce8f1c3a2