The sprite cache now supports animations and changing number of frames
When the number of frames in a spriteset changes, existing Frame arrays will not
update, but the SpriteSet (fr->sprset) and any Sprite slice will update to the
new Frame array.

There are a few places in Game (such as the load/save menu) which won't update
when frame counts change because they don't use Sprite slices yet, which could
switch from Frame ptrs to SpriteSet ptrs to update.

This also fixes crashes due to SpriteSets becoming invalid when reloading.  The
logic for when to modify the SpriteSet in-place wasn't right.

Also, only wipe the clip rect when the last reference to a clipped Frame is
deleted, not the first.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13794 7d344553-34f0-0310-a9b1-970ce8f1c3a2