Fix SpriteSet caching problems by having it include instead of inheriting from AnimationSet
Perfect example of bad object-oriented design. There was no good reason to inherit.

And updating the cache was broken: it's not actually possible to make a
SpriteSet point to a different Frame array because they share a refcount (which
was very ugly because AnimationSet also has a refcount), so it wasn't possible
to know how many of those references were to the SpriteSet and the new Frames or
should remain with the original Frames. The old code was doomed to result in
either crashing or leaking everything.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13800 7d344553-34f0-0310-a9b1-970ce8f1c3a2