Far more bounds checking when loading sprite slices from file, and in ChangeSpriteSlice
Don't showbug (log an errBug instead) when a sprite slice frame goes out of
bounds... which should now never happen. Previously, an out of range "set sprite
frame" showed a BUG message! Now it does nothing. I think it's better to not
even show a warning.

But do show a (reporterr) error when loading a slice collection with an out of
bounds spriteset.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13801 7d344553-34f0-0310-a9b1-970ce8f1c3a2