Adjust the Buy menu so a 2-line high price box doesn't overlap the items
In particular, if the length of the price text was exactly the max length of the
price box (one line), or longer (2+ lines) then the price box would grow in
height by 2 pixels due to a miscalculation, which looked quite ugly if one line.

On top of that, the items menu border box was 1 pixel taller than the info rect.
I made it the same size and position, moving it up 1 pixel.

So in total the pricebox overlaps the items menu by 3 fewer pixels, and a 2-line
price box with default box border doesn't overlap the last item.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13863 7d344553-34f0-0310-a9b1-970ce8f1c3a2