Replace VictoryState.box with explicit draw_victory_box calls
This is a debatable change, but it was actually quite confusing to figure out
when .box would be true or false, and there was a one-tick delay in
showing/hiding it in some cases.

This does not fix the victory box disappearing for a tick between each
displayed item pickup.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13871 7d344553-34f0-0310-a9b1-970ce8f1c3a2