Update "custom layout" spriteset import for variable frame sizes
It produced a spriteset with the default spriteset frame size rather than
the one the user specified, because that wasn't possible when it was written.

NOTE: it's now possible to change box border size by importing. This is
meant to be disallowed but it's not a serious problem so I'll ignore it for
the moment.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@13880 7d344553-34f0-0310-a9b1-970ce8f1c3a2