Add missing copyright notices and licenses to LICENSE.txt
(LICENSE-binary.txt updated separately)

These include notices for libraries which are included in some build we
officially distribute, and hence includes several libraries such as libopus and
libFLAC included inside SDL_mixer in certain Mac builds but which we don't use.

Also, I have added some notices which are not actually required by the license
terms (public domain, zlib and MPL) because notices as required already exist
elsewhere in the source repository.

And also I added a link to an AUTHORS.TXT file for each project if it exists.

(Backport of wip@13887)

git-svn-id: https://rpg.hamsterrepublic.com/source/rel/jocoserious@13956 7d344553-34f0-0310-a9b1-970ce8f1c3a2