Fix ItemDef constructor crash when gen() is uninitialised
This caused a crash when the 18/7/25 nightly custom.exe starts due to reading
gen(genNumElements) in a global-scope ItemDef constructor. It's hard to
reproduce because it depends on the initialisation order of globals/modules,
which can depend on OS for example.

This fixes the constructor to not crash but the deeper problem here is having
global-scope objects with constructors that depend on other globals being
initialised (see next commit).

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@14003 7d344553-34f0-0310-a9b1-970ce8f1c3a2