ItemEditor: create a blank ItemDef when switching to a new record
This happens when using the in-menu record switcher, rather than the New Item
button in the items list.

ItemEditor is the first EditorKit menu to use that ability.  The max record
variable is now increased after EditorKit.load() is called, so load() can check
for this case. Maybe in future there could be a separate overrideable
EditorKit.load_new() method but right now it seems not needed.

Also fixed prev_menu_text in the stat bonuses submenu.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@14007 7d344553-34f0-0310-a9b1-970ce8f1c3a2