Add Kaleidophone+1 to whatsnew... fix binsize.bin BUG when importing scripts during Test Game
Don't show an error when binsize.bin or fixbits.bin are written to. It looks to
me that these are reloaded correctly anyway and that nothing else would need
reloading.

I also removed a setbinsize from save_lookup1_bin that I left in accidentally
(it's redundant to the one in importscripts, which is in the more correct
location).

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@14042 7d344553-34f0-0310-a9b1-970ce8f1c3a2