Fix #1284 possible double-free when scripts double-trigger, plus byval vector bugs
The loop in run_queued_scripts failed to handle mainFibreGroup shrinking during
iteration, as I neglected to account for FB caching the iteration end point in a temp variable.

Although this happened on all OSes, we've only see this on Macs, because on Linux free()
clobbers the ScriptFibre in a way that run_queued_scripts skips over it.
There are no crash reports for this, so probably doesn't happen on Windows either.

Also, fix failing to pass mainFibreGroup around byref, which also would crash in
the right (unlikely) circumstances.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@14043 7d344553-34f0-0310-a9b1-970ce8f1c3a2