Revert unwanted music_sdl.bas changes from the web port in r13553 (9d415cdd)
These prevented debug log reporting of supported audio formats, and the modplug
settings menu, on Windows and Mac.

These happened because Blain was working from an older copy of the source and
changes I'd made in the meantime were lost. I noticed this problem at one time,
but James was the one to actually merge.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@14101 7d344553-34f0-0310-a9b1-970ce8f1c3a2