Allow Cover Children on Scroll slices and disallow on Layout and partially on Text
Although cover:all on a scroll slice is pretty silly, it makes sense to use
cover:horiz and cover:vert.

For Text slices specifically: only allow horizontal cover children, and only
if wrap is enabled. This admittedly is strange and complex.

Layout slices ignored Cover Children, but the slice editor allowed it.

(This change affects only the slice editor and set_slice_property)

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@14135 7d344553-34f0-0310-a9b1-970ce8f1c3a2