Remove Slice.Attach. It will be missed
It seemed such a nice idea that I hadn't wanted to remove it despite being
unusued for all these years, in the hope that we would actually use it some
day. However it doesn't really make sense: instead of the attach slice being
responsible for positioning but not draw order what's really needed is the
opposite. That could still be implemented with a similar system.

Instead, ScreenSlice is used instead of Parent for refreshing if Parent is NULL.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@14150 7d344553-34f0-0310-a9b1-970ce8f1c3a2