Rip apart Slice refreshing and drawing, and support recursive Cover Children
Cover Children timing has changed quite significantly, while everything else is
mostly the same.

This moves a lot of logic from DrawSliceRecurse & co to RefreshSliceTreeRecurse
where it is far easier to follow. However, I haven't changed RefreshSlice yet,
so it doesn't support covering (at all), autosorting, or much of anything
else. So this really only fixes the recursive covering FIXMEs in autotest.

I commmented out the context_stack and may removed it.

Another minor change is that DrawSlice causes a RefreshSlice on the root, which
refreshes the ancestors of the root. This doesn't seem to have an effect.

git-svn-id: https://rpg.hamsterrepublic.com/source/wip@14171 7d344553-34f0-0310-a9b1-970ce8f1c3a2